pyproject.toml
[project]
name = "devsync"
version = "0.1.0"
description = "Sync development environments between machines - git repos, secrets, configs, and more"
authors = [{name = "DevSync Contributors"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
keyring = ["keyring>=24.0.0"]
onepassword = ["onepassword-sdk>=0.1.0"]
bitwarden = ["bitwarden-sdk>=0.1.0"]

[project.scripts]
devsync = "devsync.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]